refactor: extract GraphLegend component (draggable legend with pointer-capture drag)
