+++ relfinder/src/components/graph/GraphEmptyState.vue
@@ -1,0 +1,9 @@
+<template>
+  <div class="canvas-empty">
+    <div class="empty-icon-wrap">
+      <i class="pi pi-share-alt empty-icon" />
+    </div>
+    <p class="empty-title">No graph loaded</p>
+    <p class="empty-hint">Select two entities and click <strong>Find Relationships</strong></p>
+  </div>
+</template>
